==> Analysis/1-tidy_data.Rmd <==
---
title: "Data tidying"
author: "Katie Thomas"
date: "9/21/2021"
output:
  html_document:
    keep_md: true
    code_fold: hide
    theme: flatly
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
---

<style type="text/css">

body{ /* Normal  */
      font-size: 17px;
  }
  
</style>


```{r setup, include=FALSE}

#set rmarkdown options
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, fig.height = 5, fig.width = 8, attr.output='style="max-height: 300px;"') 

#load libraries
library(cowplot)
library(plyr)
library(knitr)
library(kableExtra)
library(gridExtra)
library(evobiR)
library(geiger)
library(phylotools)
library(ape)
library(caper)
library(phytools)
library(smatr)
library(lme4)
library(tidyverse)
library(plotly)
```

# Data

## Morphological data

Morphological measurements (eye diameter and body length) were taken from a mixture of fresh and preserved specimens. In total, 459 sergestid specimens were sampled from 16 species belonging to 12 genera and two subgroups (Sergestes group and Sergia group). Measurements were recorded to ±0.01mm, but we rounded all to the nearest ±0.1mm as we were not confident in the accuracy of such fine measurements taken with calipers on a rocking ship. 

```{r import-data}

# Full specimen morphology dataset  ------

#import data
specimens_raw <- data.frame(read.csv("../Data/raw data/specimen_data.csv", header=TRUE, na.strings=c("", "NA", " ", stringsAsFactors = FALSE)))

#tidy data
specimens <- specimens_raw %>%
  mutate(Genus = gsub('\\s+', '', Genus)) %>%
  mutate(Family = str_to_title(Family)) %>%
  mutate(Subgroup = str_to_title(Subgroup)) %>%
  mutate(Genus = str_to_title(Genus)) %>%
  mutate(genus_species = as.factor(paste(Genus, Species, sep = "_"))) %>%
  mutate_if(is.numeric,round, 1)
```

```{r}

# Show sampling -----

#compile number of shrimp sampled by species
counts <-ddply(specimens, .(specimens$Subgroup, specimens$Genus, specimens$Species), nrow) 

#create table column names
names(counts) <- c("Subgroup", "Genus","Species","Sampled") 

#generate scrolling table in RMarkdown
kable(counts[ , c("Subgroup", "Genus","Species","Sampled")], caption = "Species and sampling effort for morphological data from family Sergestidae.") %>% 
                      kable_styling(full_width = F) %>% 
                      collapse_rows(columns = 1, valign = "top") %>%
                      scroll_box(height = "400px") 
```

## Trait data

Traits of interest (daytime, nighttime, and maximum reported depths, types of bioluminescent structures, habitat) were gathered for each species from published literature. Species means for eye and body size are included in this dataset.

```{r}

# Species traits  ------

#import data
traits_raw <- data.frame(read.csv("../Data/raw data/species_data.csv", header=TRUE, na.strings=c("", "NA", " ", stringsAsFactors = FALSE)))

#tidy data
traits <- traits_raw %>%
  mutate(Genus = gsub('\\s+', '', Genus)) %>%
  mutate(genus_species = as.factor(paste(Genus, Species, sep = "_"))) %>%
  mutate(Range = Day_Avg - Night_Avg) %>%
  mutate(nightbin = ifelse(Night_Avg<1000, "less","more")) %>%
  mutate(daybin = ifelse(Day_Avg<1000, "less","more"))
```

## Molecular phylogeny

We imported the sergestid molecular phylogeny generated by Golightly et al. (2021) and used _ape_ v.5.4-1 to prune the tree to the 16 species in our dataset and make the tree ultrametric (using the chronopl function with lambda set to 1).

```{r import-phylo, fig.height=5, fig.width=5}

#Import phylogeny
tree_orig <- read.tree(file = "../Data/raw data/sertestid_tree_rooted")

#Make copy of tree to manipulate
tree_edit <- tree_orig

#Remove locality tags from phylo tip labels 
n <- 3 #which _ to end pattern at
pat <- paste0('^([^_]+(?:_[^_]+){',n-1,'}).*') #pattern
tree_edit$tip.label <- sub(pat, '\\1', tree_edit$tip.label) #remove tags from tips

#Make list of taxa to drop (in tree but not in dataset)
drops <- setdiff(tree_edit$tip.label, as.character(traits$Binomial))

#Drop unwanted tips from phylogeny
tree.pruned <- drop.tip(phy = tree_edit, tip = drops) 

#Check and reorder tree
is.rooted(tree.pruned) 
is.binary.tree(tree.pruned) 
tree <- ladderize(tree.pruned) 

#Make row names of the species the phylogeny tip labels
rownames(traits) <- traits$Binomial

#Reorder species data to match phylogeny order
traits <-ReorderData(tree, traits, taxa.names="row names")

#check that phylogeny tips and data match exactly (if they match will return "OK")
name.check(tree, traits)

#make tree ultrametric
tree <- chronopl(tree, 1)

#plot tree of sampled species with original tip labels
plot.phylo(tree, #phylogeny to plot
           type = "phylogram", #how to visualize phylogeny
           show.tip.label = TRUE, #whether to show tip labels/species names
           cex = 1.0, #text size
           no.margin = TRUE, 
           use.edge.length = TRUE,
           edge.width = 2, #thickness of phylogeny branches
           label.offset = 0.1) #how far away from tips to put labels

# Make Genus_species tip labels to replace Binomial codes
sp.tips <- data.frame(old = traits$Binomial, new = traits$genus_species)

# replace tip labels in phylogeny
tree_renamed <- sub.taxa.label(tree, sp.tips)

#plot tree of sampled species with new species tip labels
plot.phylo(tree_renamed, #phylogeny to plot
           type = "phylogram", #how to visualize phylogeny
           show.tip.label = TRUE, #whether to show tip labels/species names
           cex = 1.0, #text size
           no.margin = TRUE, 
           use.edge.length = TRUE,
           edge.width = 2, #thickness of phylogeny branches
           label.offset = 0.1) #how far away from tips to put labels
```
```{r}
#remove binomial from specimen dataframe
specimens_final <- specimens %>%
  select(genus_species, Subgroup, Genus, Species, Eye_Diameter, Body_Length, Preservation) %>%
  remove_rownames()

#remove binomial from species dataframe
traits_final <- traits %>%
  select(genus_species, Subgroup, Eye_diameter, Body_length, Relative_size, Organ, Benthopelagic, Day_Avg, Night_Avg, Max_reported, Range, daybin, nightbin) %>%
  remove_rownames()
```


# Data export

We export these tidied datasets and phylogeny for use in our analyses. 

```{r}

#export tidied specimen data
write.csv(specimens_final, file = "../Data/cleaned data/specimen_data_tidy.csv", row.names = FALSE)

#export tidied trait data
write.csv(traits_final, file = "../Data/cleaned data/trait_data_tidy.csv", row.names = FALSE)

#export pruned phylogeny
write.nexus(tree_renamed, file = "../Data/cleaned data/sergestid_tree_pruned")
```
