==> Analysis/2-static_allometry.Rmd <==
---
title: "Static eye-body allometry in Sergestidae"
author: "Katie Thomas"
date: "9/21/2021"
output:
  html_document:
    keep_md: true
    code_fold: hide
    theme: flatly
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
---

<style type="text/css">

body{ /* Normal  */
      font-size: 17px;
  }
  
</style>


```{r setup, include=FALSE}

#set rmarkdown options
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, fig.height = 5, fig.width = 8, attr.output='style="max-height: 300px;"') 

#load libraries
library(cowplot)
library(plyr)
library(knitr)
library(kableExtra)
library(gridExtra)
library(evobiR)
library(geiger)
library(phylotools)
library(ape)
library(caper)
library(phytools)
library(smatr)
library(lme4)
library(tidyverse)
library(plotly)
```


# Eye-body allometry in each species

We first examine static post-metamorphosis eye diameter-body length allometry across 16 species of deep-sea sergestids. Allometric relationships were fit using standardised major axis regression via the _smatr_ v.3.4.8 package. For each species, standard model checks, model fits, and an interactive plot of data and fits are shown. Hover over the plots to see individual data values. Points are colored by the source of the sample (fresh collection or preserved in ethanol or paraformaldehyde).

Note that for 3 of the species, a large/unreasonable outlier was present. We identified these by eye and they have been removed from the data subsets and analyses, but the outlier value is noted under each species header for reference, and a second plot including the outlier as a red diamond is included. 

```{r}
#import specimen data
specimens <- data.frame(read.csv("../Data/cleaned data/specimen_data_tidy.csv", header=TRUE))
```


```{r pres-cols}
#set colors for fresh and preserved samples
col_pres <- c("ethanol" = "#a6cee3",
              "fresh" = "#b2df8a",
              "paraformaldehyde" = "#1f78b4")
```

## _Allosergestes pectinatus_

```{r apec}

# Subset data -----
apec <- specimens %>% filter(genus_species == "Allosergestes_pectinatus")

# Fit SMA model ------
sma_apec <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = apec, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_apec, which = "default",type = "o") 
plot(sma_apec, which = "residual",type = "o")
plot(sma_apec, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_apec)

#save coefficients of fits as object
cc_sma_apec <- data.frame(coef(sma_apec))

# Make plot ------
plot_apec <- ggplot(apec, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Allosergestes pectinatus") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_apec, aes(intercept = cc_sma_apec[1,1], slope = cc_sma_apec[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_apec[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_apec[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

ggplotly(plot_apec)
```

## _Allosergestes sargassi_

Note: this species has a big outlier that was removed for analyses and plots (eye 0.7, body 13.0). 

```{r asar}

# Subset data -----
#note: outlier removed
asar <- specimens %>% 
  filter(genus_species == "Allosergestes_sargassi") %>%
  filter(!(Eye_Diameter == round(0.67, 1) & Body_Length == round(12.97,1)))

# Fit SMA model ------
sma_asar <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = asar, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_asar, which = "default",type = "o") 
plot(sma_asar, which = "residual",type = "o")
plot(sma_asar, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_asar)

#save coefficients of fits as object
cc_sma_asar <- data.frame(coef(sma_asar))

# Make plot ------
plot_asar <- ggplot(asar, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Allosergestes sargassi") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_asar, aes(intercept = cc_sma_asar[1,1], slope = cc_sma_asar[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_asar[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_asar[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

#plot with outlier shown
plot_asar_out <- plot_asar +
  geom_point(aes(y=round(0.67, 1), x=round(12.97,1)), colour="red", pch = 18) +
  ggtitle("Allosergestes sargassi + outlier")
  
#plot
ggplotly(plot_asar)

#plot with outlier
ggplotly(plot_asar_out)
```

## _Challengerosergia hansjacobi_

```{r chan}

# Subset data -----
chan <- specimens %>% filter(genus_species == "Challengerosergia_hansjacobi")

# Fit SMA model ------
sma_chan <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = chan, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_chan, which = "default",type = "o") 
plot(sma_chan, which = "residual",type = "o")
plot(sma_chan, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_chan)

#save coefficients of fits as object
cc_sma_chan <- data.frame(coef(sma_chan))

# Make plot ------
plot_chan <- ggplot(chan, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Challengerosergia hansjacobi") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_chan, aes(intercept = cc_sma_chan[1,1], slope = cc_sma_chan[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_chan[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_chan[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

ggplotly(plot_chan)
```

## _Challengerosergia talismani_

```{r ctal}

# Subset data -----
ctal <- specimens %>% filter(genus_species == "Challengerosergia_talismani")

# Fit SMA model ------
sma_ctal <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = ctal, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_ctal, which = "default",type = "o") 
plot(sma_ctal, which = "residual",type = "o")
plot(sma_ctal, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_ctal)

#save coefficients of fits as object
cc_sma_ctal <- data.frame(coef(sma_ctal))

# Make plot ------
plot_ctal <- ggplot(ctal, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Challengerosergia talismani") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_ctal, aes(intercept = cc_sma_ctal[1,1], slope = cc_sma_ctal[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_ctal[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_ctal[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

ggplotly(plot_ctal)
```

## _Deosergestes corniculum_

```{r dcor}

# Subset data -----
dcor <- specimens %>% filter(genus_species == "Deosergestes_corniculum")

# Fit SMA model ------
sma_dcor <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = dcor, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_dcor, which = "default",type = "o") 
plot(sma_dcor, which = "residual",type = "o")
plot(sma_dcor, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_dcor)

#save coefficients of fits as object
cc_sma_dcor <- data.frame(coef(sma_dcor))

# Make plot ------
plot_dcor <- ggplot(dcor, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Deosergestes corniculum") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_dcor, aes(intercept = cc_sma_dcor[1,1], slope = cc_sma_dcor[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_dcor[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_dcor[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

ggplotly(plot_dcor)
```

## _Deosergestes henseni_

```{r dhen}

# Subset data -----
dhen <- specimens %>% filter(genus_species == "Deosergestes_henseni")

# Fit SMA model ------
sma_dhen <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = dhen, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_dhen, which = "default",type = "o") 
plot(sma_dhen, which = "residual",type = "o")
plot(sma_dhen, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_dhen)

#save coefficients of fits as object
cc_sma_dhen <- data.frame(coef(sma_dhen))

# Make plot ------
plot_dhen <- ggplot(dhen, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Deosergestes henseni") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_dhen, aes(intercept = cc_sma_dhen[1,1], slope = cc_sma_dhen[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_dhen[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_dhen[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

ggplotly(plot_dhen)
```

## _Eusergestes arcticus_ 

```{r euar}

# Subset data -----
euar <- specimens %>% filter(genus_species == "Eusergestes_arcticus")

# Fit SMA model ------
sma_euar <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = euar, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_euar, which = "default",type = "o") 
plot(sma_euar, which = "residual",type = "o")
plot(sma_euar, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_euar)

#save coefficients of fits as object
cc_sma_euar <- data.frame(coef(sma_euar))

# Make plot ------
plot_euar <- ggplot(euar, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Eusergestes arcticus") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_euar, aes(intercept = cc_sma_euar[1,1], slope = cc_sma_euar[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_euar[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_euar[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

ggplotly(plot_euar)
```

## _Gardinerosergia splendens_

Note: this species has a big outlier that was removed for analyses and plots (eye 1.7, body 22.0). 

```{r gspl}

# Subset data -----
gspl <- specimens %>% 
  filter(genus_species == "Gardinerosergia_splendens") %>%
  filter(!(Eye_Diameter == round(1.71, 1) & Body_Length == round(21.98,1)))

# Fit SMA model ------
sma_gspl <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = gspl, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_gspl, which = "default",type = "o") 
plot(sma_gspl, which = "residual",type = "o")
plot(sma_gspl, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_gspl)

#save coefficients of fits as object
cc_sma_gspl <- data.frame(coef(sma_gspl))

# Make plot ------
plot_gspl <- ggplot(gspl, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Gardinerosergia splendens") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_gspl, aes(intercept = cc_sma_gspl[1,1], slope = cc_sma_gspl[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_gspl[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_gspl[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

#plot with outlier shown
plot_gspl_out <- plot_gspl +
  geom_point(aes(y=round(1.71,1), x=round(21.98,1)), colour="red", pch = 18) +
  ggtitle("Gardinerosergia splendens + outlier")
  
#plot
ggplotly(plot_gspl)

#plot with outlier
ggplotly(plot_gspl_out)
```

## _Neosergestes edwardsii_

Note: this species has a big outlier that was removed for analyses and plots (eye 0.9, body 0.7). Also note that for this species, sample size was insufficient to estimate eye-body allometry (N = 6, p = 0.57 for correlation between eye size and body size). 

```{r nedw}

# Subset data -----
nedw <- specimens %>% 
  filter(genus_species == "Neosergestes_edwardsii") %>%
  filter(!(Eye_Diameter == round(0.85,1) & Body_Length == round(0.72,1)))

# Fit SMA model ------
sma_nedw <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = nedw, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_nedw, which = "default",type = "o") 
plot(sma_nedw, which = "residual",type = "o")
plot(sma_nedw, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_nedw)

#save coefficients of fits as object
cc_sma_nedw <- data.frame(coef(sma_nedw))

# Make plot ------
plot_nedw <- ggplot(nedw, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Neosergestes edwardsii") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) 

#plot with outlier shown
plot_nedw_out <- plot_nedw +
  geom_point(aes(x=round(0.85,1), y=round(0.72,1)), colour="red", pch = 18) +
  ggtitle("Neosergestes edwardsii + outlier")
  
#plot
ggplotly(plot_nedw)

#plot with outlier
ggplotly(plot_nedw_out)
```

## _Parasergestes armatus_

```{r parm}

# Subset data -----
parm <- specimens %>% filter(genus_species == "Parasergestes_armatus")

# Fit SMA model ------
sma_parm <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = parm, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_parm, which = "default",type = "o") 
plot(sma_parm, which = "residual",type = "o")
plot(sma_parm, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_parm)

#save coefficients of fits as object
cc_sma_parm <- data.frame(coef(sma_parm))

# Make plot ------
plot_parm <- ggplot(parm, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Parasergestes armatus") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_parm, aes(intercept = cc_sma_parm[1,1], slope = cc_sma_parm[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_parm[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_parm[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

ggplotly(plot_parm)
```

## _Parasergestes vigilax_

```{r pvig}

# Subset data -----
pvig <- specimens %>% filter(genus_species == "Parasergestes_vigilax")

# Fit SMA model ------
sma_pvig <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = pvig, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_pvig, which = "default",type = "o") 
plot(sma_pvig, which = "residual",type = "o")
plot(sma_pvig, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_pvig)

#save coefficients of fits as object
cc_sma_pvig <- data.frame(coef(sma_pvig))

# Make plot ------
plot_pvig <- ggplot(pvig, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Parasergestes vigilax") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_pvig, aes(intercept = cc_sma_pvig[1,1], slope = cc_sma_pvig[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_pvig[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_pvig[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

ggplotly(plot_pvig)
```

## _Phorcosergia grandis_

```{r pgra}

# Subset data -----
pgra <- specimens %>% filter(genus_species == "Phorcosergia_grandis")

# Fit SMA model ------
sma_pgra <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = pgra, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_pgra, which = "default",type = "o") 
plot(sma_pgra, which = "residual",type = "o")
plot(sma_pgra, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_pgra)

#save coefficients of fits as object
cc_sma_pgra <- data.frame(coef(sma_pgra))

# Make plot ------
plot_pgra <- ggplot(pgra, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Phorcosergia grandis") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_pgra, aes(intercept = cc_sma_pgra[1,1], slope = cc_sma_pgra[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_pgra[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_pgra[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

ggplotly(plot_pgra)
```

## _Robustosergia robusta_

```{r rrob}

# Subset data -----
rrob <- specimens %>% filter(genus_species == "Robustosergia_robusta")

# Fit SMA model ------
sma_rrob <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = rrob, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_rrob, which = "default",type = "o") 
plot(sma_rrob, which = "residual",type = "o")
plot(sma_rrob, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_rrob)

#save coefficients of fits as object
cc_sma_rrob <- data.frame(coef(sma_rrob))

# Make plot ------
plot_rrob <- ggplot(rrob, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Robustosergia robusta") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_rrob, aes(intercept = cc_sma_rrob[1,1], slope = cc_sma_rrob[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_rrob[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_rrob[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

ggplotly(plot_rrob)
```

## _Robustosergia regalis_

```{r rreg}

# Subset data -----
rreg <- specimens %>% filter(genus_species == "Robustosergia_regalis")

# Fit SMA model ------
sma_rreg <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = rreg, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_rreg, which = "default",type = "o") 
plot(sma_rreg, which = "residual",type = "o")
plot(sma_rreg, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_rreg)

#save coefficients of fits as object
cc_sma_rreg <- data.frame(coef(sma_rreg))

# Make plot ------
plot_rreg <- ggplot(rreg, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Robustosergia regalis") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_rreg, aes(intercept = cc_sma_rreg[1,1], slope = cc_sma_rreg[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_rreg[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_rreg[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

ggplotly(plot_rreg)
```

## _Sergestes atlanticus_

```{r satl}

# Subset data -----
satl <- specimens %>% filter(genus_species == "Sergestes_atlanticus")

# Fit SMA model ------
sma_satl <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = satl, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_satl, which = "default",type = "o") 
plot(sma_satl, which = "residual",type = "o")
plot(sma_satl, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_satl)

#save coefficients of fits as object
cc_sma_satl <- data.frame(coef(sma_satl))

# Make plot ------
plot_satl <- ggplot(satl, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Sergestes atlanticus") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_satl, aes(intercept = cc_sma_satl[1,1], slope = cc_sma_satl[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_satl[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_satl[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

ggplotly(plot_satl)
```

## _Sergia tenuiremis_

```{r sten}

# Subset data -----
sten <- specimens %>% filter(genus_species == "Sergia_tenuiremis")

# Fit SMA model ------
sma_sten <- sma(formula = Eye_Diameter ~ Body_Length, 
            data = sten, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            alpha = 0.05)

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_sten, which = "default",type = "o") 
plot(sma_sten, which = "residual",type = "o")
plot(sma_sten, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_sten)

#save coefficients of fits as object
cc_sma_sten <- data.frame(coef(sma_sten))

# Make plot ------
plot_sten <- ggplot(sten, aes(x = Body_Length, y = Eye_Diameter, col = Preservation)) + 
  geom_point(size = 2, alpha = 0.8) + 
  scale_color_manual(values = col_pres, name = "Sample state", breaks = c("ethanol", "fresh", "paraformaldehyde")) +
  scale_x_log10(name = "Body length (mm)") + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)") + #makes y axis log scale and named
  ggtitle("Sergia tenuiremis") + 
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), plot.title = element_text(face = "italic")) +
  geom_abline(data = cc_sma_sten, aes(intercept = cc_sma_sten[1,1], slope = cc_sma_sten[2,1])) +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 1.5, nudge_x = 4, label = paste("b =", round(cc_sma_sten[2,1], digits = 2), sep = " "), size = 3.5, col = "black") +
  geom_text(x = -Inf, y = Inf, hjust = -0.05, vjust = 3, label = paste("a =", round(cc_sma_sten[1,1], digits = 2), sep = " "), size = 3.5, col = "black")

ggplotly(plot_sten)
```

## Figure for static eye-body scaling across sergestid species

```{r figure-static, fig.width = 10, fig.height = 12, results = "hide"}

#make figure panels
fig.a <- plot_dcor + theme(legend.position = "none", axis.title.x = element_blank())
fig.b <- plot_dhen + theme(legend.position = "none", axis.title.x = element_blank(), axis.title.y = element_blank())
fig.c <- plot_apec + theme(legend.position = "none", axis.title.x = element_blank(), axis.title.y = element_blank())
fig.d <- plot_asar + theme(legend.position = "none", axis.title.x = element_blank())
fig.e <- plot_satl + theme(legend.position = "none", axis.title.x = element_blank(), axis.title.y = element_blank())
fig.f <- plot_pvig + theme(legend.position = "none", axis.title.x = element_blank(), axis.title.y = element_blank())
fig.g <- plot_parm + theme(legend.position = "none", axis.title.x = element_blank())
fig.h <- plot_euar + theme(legend.position = "none", axis.title.x = element_blank(), axis.title.y = element_blank())
fig.i <- plot_gspl + theme(legend.position = "none", axis.title.x = element_blank(), axis.title.y = element_blank())
fig.j <- plot_rreg + theme(legend.position = "none", axis.title.x = element_blank())
fig.k <- plot_rrob + theme(legend.position = "none", axis.title.x = element_blank(), axis.title.y = element_blank())
fig.l <- plot_pgra + theme(legend.position = "none", axis.title.x = element_blank(), axis.title.y = element_blank())
fig.m <- plot_sten + theme(legend.position = "none")
fig.n <- plot_ctal  + theme(legend.position = "none", axis.title.y = element_blank())
fig.o <- plot_chan + theme(legend.position = "none", axis.title.y = element_blank())

# arrange panels in figure
plots <- plot_grid(fig.a, fig.b, fig.c, fig.d, fig.e, fig.f, fig.g, fig.h, fig.i, fig.j, fig.k, fig.l, fig.m, fig.n, fig.o, #list of plots to arrange in grid
           align = 'vh', #align horizontally and vertically
           axis = 'lb',
           labels = c("A", "B", "C", "D", "E", "F", "G", "H", "I","J", "K", "L", "M", "N", "O"), #panel labels for figure
           hjust = -1, #adjustment for panel labels
           nrow = 5) #number of rows in grids

#extract legend
leg <- get_legend(fig.e + theme(legend.position="bottom"))


#view figure
plot_grid(plots, leg, nrow = 2, rel_heights = c(1, .01), align = 'vh')

#export figure
#png("../Figures/dev-allometry.png", width = 10, height = 12, units = "in", res = 500)
pdf("../Figures/Fig-S1.pdf", width = 9, height = 12)
plot_grid(plots, leg, nrow = 2, rel_heights = c(1, .05), align = 'vh')
dev.off()
```

# Test whether species differ in static allometry

We've looked at how eyes scale with body size within each species, but are there significant differences in static allometry across species? Here we use two approaches: first, we compare linear mixed models to see which best describes our data. Then, we look at pairwise comparisons of SMA regressions among species to see which are driving these differences. 

## Linear mixed model to test whether slopes vary across species

Here, we use linear mixed models in _lme4_ v.1.1.25 to test whether species show differences in static eye-body allometric slopes. We fit 1) a constant-slope, variable intercept model to our eye and body size data with species as a random effect and 2) a variable slope model with the same effects, and then compared model fits by AIC scores. Both models are fit using reduced maximum likelihood (REML), as this approach is better than maximum likelihood (ML) for detecting differences in random effects among models, which is what we are testing here (ML is better for fixed effects). 

First we removed the three outliers and the species with insufficient sampling (_N. edwardsii_) from our full specimen dataset, then ran each model. 

```{r prep-data}
#Remove outliers from specimen dataset
specimens_ed <- specimens %>%
  filter(!(genus_species == "Allosergestes_sargassi" & Eye_Diameter == 0.7 & Body_Length == 13.0)) %>% #(0.67, 12.97 raw)
  filter(!(genus_species == "Gardinerosergia_splendens" & Eye_Diameter == 1.7 & Body_Length == 22.0)) %>% #(1.71, 21.98 raw)
  filter(!(genus_species == "Neosergestes_edwardsii" & Eye_Diameter == 0.8 & Body_Length == 0.7)) #(0.85, 0.72 raw)

#Export data without outliers
write.csv(specimens_ed, file = "../Data/cleaned data/specimen_data_tidy_no-outliers.csv", row.names = FALSE)

#Remove species with insufficient sampling
specimens_test <- specimens_ed %>%
  filter(genus_species != "Neosergestes_edwardsii") %>%
  mutate(species_text = as.factor(paste(Genus, Species, sep = " "))) # Add labels for species text

```

### Fixed slope (variable intercepts) model

```{r fixed-slope}

# variable intercepts model (species can have different intercepts but share mean slope) -------
lme_fixed <- lmer(log10(Eye_Diameter) ~ log10(Body_Length) + (1|genus_species), 
                  data = specimens_test)

# summary 
summary(lme_fixed)

# intercepts (variable) and slope (constant)
coef(lme_fixed)

# AIC
AIC(lme_fixed)

#fixed effects
fixef(lme_fixed)

#confidence intervals
## if these overlap zero, we do not have good support for significant effects
## if .sig01, which is our estimate of the variability in the random effect is very large and very widely defined, t indicates we may have a lack of precision between our groups - either because the group effect is small between groups, we have too few groups to get a more precise estimate, we have too few units within each group, or a combination of all of the above.
confint(lme_fixed, level = 0.95)

```

### Variable slopes model 

```{r variable-slope}
# variable slopes model (species can have different slopes/interaction effect between body size and species on eye size) -------
lme_variable <- lmer(log10(Eye_Diameter) ~ log10(Body_Length) + (1 + log10(Body_Length) | genus_species), 
                     data = specimens_test)

# summary
summary(lme_variable)

# intercepts (variable) and slope (constant)
coef(lme_variable)


# AIC
AIC(lme_variable)

#fixed effects
fixef(lme_variable)

#confidence intervals
## if these overlap zero, we do not have good support for significant effects
## if .sig01, which is our estimate of the variability in the random effect is very large and very widely defined, t indicates we may have a lack of precision between our groups - either because the group effect is small between groups, we have too few groups to get a more precise estimate, we have too few units within each group, or a combination of all of the above.
confint(lme_variable, level = 0.95)
```

Note that the parameter estimates for slopes and intercepts for species groups are a bit different than those we estimated via OLS. That's because here, the model is looking across all the data for all the species, and finding a common mean slope/intercept as well as the group effects. 

### Compare models

Then we can compare the two models via AIC and log likelihood. 

```{r}

#model AIC comparison
AIC(lme_fixed, lme_variable)

#log likelihood comparison
anova(lme_fixed, lme_variable, refit=FALSE)
```

Here we find by comparing AIC scores and log likelihoods that the model with variable slopes is better supported than the one with fixed slopes. We have evidence here that species differ significantly in the slopes of their static eye-body allometries rather than following a shared growth trajectory. 


## Test for significant effect of preservation as a covariate

Here, we add preservation method as a fixed effect into our linear mixed models of eye size vs. body size with species as a random effect. We have already selected the variable slopes model, so here we use a maximum likelihood approach (ML) to fit models that 1) don't include preservation as a covariate and 2) do include preservation as a covariate and then compare model fits. We include preservation as a fixed rather than a random effect because random effects should have more than 5 levels, and preservation type has only 3. 

### Model without preservation type

```{r no-pres}

# variable slopes model (species can have different slopes/interaction effect between body size and species on eye size) -------
lme_nopres <- lmer(log10(Eye_Diameter) ~ log10(Body_Length) + (1 + log10(Body_Length) | genus_species), 
                   data = specimens_test,
                   REML = FALSE) #uses ML instead of REML

# summary
summary(lme_nopres)

# intercepts (variable) and slope (constant)
coef(lme_nopres)

# AIC
AIC(lme_nopres)

#fixed effects
fixef(lme_nopres)

#confidence intervals
## if these overlap zero, we do not have good support for significant effects
## if .sig01, which is our estimate of the variability in the random effect is very large and very widely defined, t indicates we may have a lack of precision between our groups - either because the group effect is small between groups, we have too few groups to get a more precise estimate, we have too few units within each group, or a combination of all of the above.
confint(lme_nopres, level = 0.95)
```

### Model including preservation type

```{r with-pres}

#reorder levels for preservation so that contrasts will be compared to fresh samples
specimens_test$Preservation <- factor(specimens_test$Preservation, 
                                      levels = c("fresh", "ethanol", "paraformaldehyde"))

# variable slopes model (species can have different slopes/interaction effect between body size and species on eye size) -------
lme_withpres <- lmer(log10(Eye_Diameter) ~ log10(Body_Length) + Preservation + (1 + log10(Body_Length) | genus_species), 
                   data = specimens_test,
                   REML = FALSE) #uses ML instead of REML

# summary
summary(lme_withpres)

# AIC
AIC(lme_withpres)

#fixed effects
fixef(lme_withpres)

#confidence intervals
## if these overlap zero, we do not have good support for significant effects
## if .sig01, which is our estimate of the variability in the random effect is very large and very widely defined, t indicates we may have a lack of precision between our groups - either because the group effect is small between groups, we have too few groups to get a more precise estimate, we have too few units within each group, or a combination of all of the above.
confint(lme_withpres, level = 0.95)
```

### Compare models

Then we can compare the two models via AIC and log likelihood.

```{r}

#model AIC comparison
AIC(lme_nopres, lme_withpres)

#log likelihood comparison
anova(lme_nopres, lme_withpres)
```

Here we see that the model that includes preservation type as a covariate is significantly better supported. Ethanol preservation has a negative effect on intercept (eye size compared to body size), but paraformaldehyde has a positive effect. Note also that these effects are very small, so while they are significantly affecting our models, they seem to have minor effects.

## Linear SMA regression for pairwise comparisons among species slopes

The _smatr_ package also allows you to test whether groups exhibit significant differences in allometric scaling (slopes or intercepts). Below, we again use standardized major axis regression to model the scaling of eye diameter with body length across all specimens measured and test whether we see significant differences in static allometry across species. 

```{r species-smas}

# Model for static allometry across specimens with species and species*body size interaction as covariates
sma_species <- sma(formula = Eye_Diameter ~ Body_Length * genus_species, 
            data = specimens_test, 
            log = "xy", #sets both x and y variables as logged
            method="SMA", #defines SMA as model
            multcomp = TRUE, # adds pairwise comparisons between groups
            multcompmethod = "adjusted", # adjusts p-value for multiple comparisons
            alpha = 0.05) 

#plot fit, residuals, qq
par(mfrow = c(2,2)) #make plot window 2x2
plot(sma_species, which = "default",type = "o") 
plot(sma_species,which = "residual",type = "o")
plot(sma_species, which = "qq", type = "o")
par(mfrow = c(1,1)) 

#view fits
summary(sma_species)

#save coefficients of fits as object
cc_species <- data.frame(coef(sma_species))
```

Among 15 sergestid species (450 specimens), species significantly differ in eye-body allometric scaling (p < 0.0001). Pairwise comparisons show which species differ significantly in slope and which do not.

Here we plot the static allometry of all the species with sufficient sampling (n = 15). It's interactive, so you can click on different species in the legend to hide or show those points.

```{r species-fig, fig.width=9, fig.height=9}

#reorder factor levels for figure legend (phylo order)
specimens_test$species_text <- factor(specimens_test$species_text, 
                                      levels = c("Deosergestes corniculum",
                                                 "Deosergestes henseni",
                                                 "Allosergestes pectinatus",
                                                 "Allosergestes sargassi",
                                                 "Sergestes atlanticus",
                                                 "Parasergestes vigilax",
                                                 "Parasergestes armatus",
                                                 "Eusergestes arcticus",
                                                 "Gardinerosergia splendens",
                                                 "Robustosergia regalis",
                                                 "Robustosergia robusta",
                                                 "Phorcosergia grandis",
                                                 "Sergia tenuiremis",
                                                 "Challengerosergia talismani",
                                                 "Challengerosergia hansjacobi"))

#make shape pallette
shapes.sp <- c("Deosergestes corniculum" = 21,
              "Deosergestes henseni" = 22, 
              "Allosergestes pectinatus" = 23, 
              "Allosergestes sargassi" = 24,
              "Sergestes atlanticus" = 25, 
              "Parasergestes vigilax" = 21, 
              "Parasergestes armatus" = 22,
              "Eusergestes arcticus" = 23, 
              "Gardinerosergia splendens" = 24, 
              "Robustosergia regalis" = 25, 
              "Robustosergia robusta" = 21, 
              "Phorcosergia grandis" = 22,
              "Sergia tenuiremis" = 23,
              "Challengerosergia talismani" = 24,
              "Challengerosergia hansjacobi" = 25)

#rainbow palette
cols.sp <- c("Deosergestes corniculum" = "#16ABA8",
              "Deosergestes henseni" = "#1f78b4",
              "Allosergestes pectinatus" = "#3acf75",
              "Allosergestes sargassi" = "#33a02c",
              "Sergestes atlanticus" = "#fb9a99",
              "Parasergestes vigilax" = "#e31a1c",
              "Parasergestes armatus" = "#fdbf6f",
              "Eusergestes arcticus" = "#ff7f00",
              "Gardinerosergia splendens" = "#cab2d6",
              "Robustosergia regalis" = "#6a3d9a",
              "Robustosergia robusta" = "gray31",
              "Phorcosergia grandis" = "burlywood4",
              "Sergia tenuiremis" = "black",
              "Challengerosergia talismani" = "maroon1",
              "Challengerosergia hansjacobi" = "orchid1")

#sergia/sergestes green purple pallette
cols.sp <- c(#Sergestes group
              "Deosergestes corniculum" = "#512E5F",
              "Deosergestes henseni" = "#633974",
              "Allosergestes pectinatus" = "#76448A",
              "Allosergestes sargassi" = "#884EA0",
              "Sergestes atlanticus" = "#9B59B6",
              #"Neosergestes edwardsii" = "#AF7AC5"
              "Parasergestes vigilax" = "#C39BD3",
              "Parasergestes armatus" = "#D7BDE2",
              "Eusergestes arcticus" = "#EBDEF0",
             #Sergia group
              "Gardinerosergia splendens" = "#D4EFDF",
              "Robustosergia regalis" = "#A9DFBF",
              "Robustosergia robusta" = "#52BE80",
              "Phorcosergia grandis" = "#27AE60",
              "Sergia tenuiremis" = "#1E8449",
              "Challengerosergia talismani" = "#196F3D",
              "Challengerosergia hansjacobi" = "#145A32")


#set s limits by species for plots
limits <- specimens_test %>%
  group_by(genus_species) %>%
  summarise(max = max(Body_Length, na.rm = TRUE),
            min = min(Body_Length, na.rm = TRUE))

#make plot
plot_species <- ggplot(specimens_test, aes(x = Body_Length, y = Eye_Diameter, color =  species_text, shape = species_text, fill = species_text)) + 
  geom_point(size = 2, alpha = 0.4) + 
  scale_shape_manual(values = shapes.sp, name = "Species") + 
  scale_color_manual(values = cols.sp, name = "Species") +
  scale_fill_manual(values = cols.sp, name = "Species") +
  scale_x_log10(name = "Body length (mm)", breaks = c(15,20,30,50,100)) + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)", breaks = c(0.2,0.3,0.5,1, 2.6)) + #makes y axis log scale and named
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), legend.text = element_text(face = "italic")) +
    geom_segment(aes(x = limits$min[which(limits$genus_species == "Allosergestes_pectinatus")], xend = limits$max[which(limits$genus_species == "Allosergestes_pectinatus")], y = 10^(log10(limits$min[which(limits$genus_species == "Allosergestes_pectinatus")])*cc_species["Allosergestes_pectinatus", "slope"] + cc_species["Allosergestes_pectinatus", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Allosergestes_pectinatus")])*cc_species["Allosergestes_pectinatus", "slope"] + cc_species["Allosergestes_pectinatus", "elevation"])), color = cols.sp["Allosergestes pectinatus"]) + #Rana temporaria adult scaling
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Allosergestes_sargassi")], xend = limits$max[which(limits$genus_species == "Allosergestes_sargassi")], y = 10^(log10(limits$min[which(limits$genus_species == "Allosergestes_sargassi")])*cc_species["Allosergestes_sargassi", "slope"] + cc_species["Allosergestes_sargassi", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Allosergestes_sargassi")])*cc_species["Allosergestes_sargassi", "slope"] + cc_species["Allosergestes_sargassi", "elevation"])), color = cols.sp["Allosergestes sargassi"]) + #Allosergestes_sargassi adult scaling
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Challengerosergia_hansjacobi")], xend = limits$max[which(limits$genus_species == "Challengerosergia_hansjacobi")], y = 10^(log10(limits$min[which(limits$genus_species == "Challengerosergia_hansjacobi")])*cc_species["Challengerosergia_hansjacobi", "slope"] + cc_species["Challengerosergia_hansjacobi", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Challengerosergia_hansjacobi")])*cc_species["Challengerosergia_hansjacobi", "slope"] + cc_species["Challengerosergia_hansjacobi", "elevation"])), color = cols.sp["Challengerosergia hansjacobi"]) + #Challengerosergia_hansjacobi adult scaling
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Challengerosergia_talismani")], xend = limits$max[which(limits$genus_species == "Challengerosergia_talismani")], y = 10^(log10(limits$min[which(limits$genus_species == "Challengerosergia_talismani")])*cc_species["Challengerosergia_talismani", "slope"] + cc_species["Challengerosergia_talismani", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Challengerosergia_talismani")])*cc_species["Challengerosergia_talismani", "slope"] + cc_species["Challengerosergia_talismani", "elevation"])), color = cols.sp["Challengerosergia talismani"]) + #Challengerosergia_talismani adult scaling
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Deosergestes_corniculum")], xend = limits$max[which(limits$genus_species == "Deosergestes_corniculum")], y = 10^(log10(limits$min[which(limits$genus_species == "Deosergestes_corniculum")])*cc_species["Deosergestes_corniculum", "slope"] + cc_species["Deosergestes_corniculum", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Deosergestes_corniculum")])*cc_species["Deosergestes_corniculum", "slope"] + cc_species["Deosergestes_corniculum", "elevation"])), color = cols.sp["Deosergestes corniculum"]) + #Deosergestes_corniculum adult scaling
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Deosergestes_henseni")], xend = limits$max[which(limits$genus_species == "Deosergestes_henseni")], y = 10^(log10(limits$min[which(limits$genus_species == "Deosergestes_henseni")])*cc_species["Deosergestes_henseni", "slope"] + cc_species["Deosergestes_henseni", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Deosergestes_henseni")])*cc_species["Deosergestes_henseni", "slope"] + cc_species["Deosergestes_henseni", "elevation"])), color = cols.sp["Deosergestes henseni"]) + #Deosergestes henseni adult scaling
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Gardinerosergia_splendens")], xend = limits$max[which(limits$genus_species == "Gardinerosergia_splendens")], y = 10^(log10(limits$min[which(limits$genus_species == "Gardinerosergia_splendens")])*cc_species["Gardinerosergia_splendens", "slope"] + cc_species["Gardinerosergia_splendens", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Gardinerosergia_splendens")])*cc_species["Gardinerosergia_splendens", "slope"] + cc_species["Gardinerosergia_splendens", "elevation"])), color = cols.sp["Gardinerosergia splendens"]) + #Gardinerosergia splendens adult scaling
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Parasergestes_armatus")], xend = limits$max[which(limits$genus_species == "Parasergestes_armatus")], y = 10^(log10(limits$min[which(limits$genus_species == "Parasergestes_armatus")])*cc_species["Parasergestes_armatus", "slope"] + cc_species["Parasergestes_armatus", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Parasergestes_armatus")])*cc_species["Parasergestes_armatus", "slope"] + cc_species["Parasergestes_armatus", "elevation"])), color = cols.sp["Parasergestes armatus"]) + #Parasergestes armatus adult scaling
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Parasergestes_vigilax")], xend = limits$max[which(limits$genus_species == "Parasergestes_vigilax")], y = 10^(log10(limits$min[which(limits$genus_species == "Parasergestes_vigilax")])*cc_species["Parasergestes_vigilax", "slope"] + cc_species["Parasergestes_vigilax", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Parasergestes_vigilax")])*cc_species["Parasergestes_vigilax", "slope"] + cc_species["Parasergestes_vigilax", "elevation"])), color = cols.sp["Parasergestes vigilax"]) + #Parasergestes vigilax adult scaling
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Phorcosergia_grandis")], xend = limits$max[which(limits$genus_species == "Phorcosergia_grandis")], y = 10^(log10(limits$min[which(limits$genus_species == "Phorcosergia_grandis")])*cc_species["Phorcosergia_grandis", "slope"] + cc_species["Phorcosergia_grandis", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Phorcosergia_grandis")])*cc_species["Phorcosergia_grandis", "slope"] + cc_species["Phorcosergia_grandis", "elevation"])), color = cols.sp["Phorcosergia grandis"]) + #Phorcosergia grandis adult scaling
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Robustosergia_robusta")], xend = limits$max[which(limits$genus_species == "Robustosergia_robusta")], y = 10^(log10(limits$min[which(limits$genus_species == "Robustosergia_robusta")])*cc_species["Robustosergia_robusta", "slope"] + cc_species["Robustosergia_robusta", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Robustosergia_robusta")])*cc_species["Robustosergia_robusta", "slope"] + cc_species["Robustosergia_robusta", "elevation"])), color = cols.sp["Robustosergia robusta"]) + #Robustosergia robusta adult scaling
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Robustosergia_regalis")], xend = limits$max[which(limits$genus_species == "Robustosergia_regalis")], y = 10^(log10(limits$min[which(limits$genus_species == "Robustosergia_regalis")])*cc_species["Robustosergia_regalis", "slope"] + cc_species["Robustosergia_regalis", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Robustosergia_regalis")])*cc_species["Robustosergia_regalis", "slope"] + cc_species["Robustosergia_regalis", "elevation"])), color = cols.sp["Robustosergia regalis"]) +
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Sergestes_atlanticus")], xend = limits$max[which(limits$genus_species == "Sergestes_atlanticus")], y = 10^(log10(limits$min[which(limits$genus_species == "Sergestes_atlanticus")])*cc_species["Sergestes_atlanticus", "slope"] + cc_species["Sergestes_atlanticus", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Sergestes_atlanticus")])*cc_species["Sergestes_atlanticus", "slope"] + cc_species["Sergestes_atlanticus", "elevation"])), color = cols.sp["Sergestes atlanticus"]) +
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Sergia_tenuiremis")], xend = limits$max[which(limits$genus_species == "Sergia_tenuiremis")], y = 10^(log10(limits$min[which(limits$genus_species == "Sergia_tenuiremis")])*cc_species["Sergia_tenuiremis", "slope"] + cc_species["Sergia_tenuiremis", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Sergia_tenuiremis")])*cc_species["Sergia_tenuiremis", "slope"] + cc_species["Sergia_tenuiremis", "elevation"])), color = cols.sp["Sergia tenuiremis"])+
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Eusergestes_arcticus")], xend = limits$max[which(limits$genus_species == "Eusergestes_arcticus")], y = 10^(log10(limits$min[which(limits$genus_species == "Eusergestes_arcticus")])*cc_species["Eusergestes_arcticus", "slope"] + cc_species["Eusergestes_arcticus", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Eusergestes_arcticus")])*cc_species["Eusergestes_arcticus", "slope"] + cc_species["Eusergestes_arcticus", "elevation"])), color = cols.sp["Eusergestes arcticus"])+
  theme(legend.text = element_text(face = "italic"))


#interactive plot
ggplotly(plot_species)
```

### Figure 1 revision

Here we again plot the static allometry of all the species with sufficient sampling (n = 15). Here per reviewer suggestions, we split the data into two panels - one for the -sergestes group and one for the -sergia group. 

First we make a panel showing the -sergestes data, with the -sergia regressions in a light gray for reference. 
```{r species-fig1a, fig.width=9, fig.height=9}

#reorder factor levels for figure legend (phylo order)
specimens_test$species_text <- factor(specimens_test$species_text, 
                                      levels = c("Deosergestes corniculum",
                                                 "Deosergestes henseni",
                                                 "Allosergestes pectinatus",
                                                 "Allosergestes sargassi",
                                                 "Sergestes atlanticus",
                                                 "Parasergestes vigilax",
                                                 "Parasergestes armatus",
                                                 "Eusergestes arcticus",
                                                 "Gardinerosergia splendens",
                                                 "Robustosergia regalis",
                                                 "Robustosergia robusta",
                                                 "Phorcosergia grandis",
                                                 "Sergia tenuiremis",
                                                 "Challengerosergia talismani",
                                                 "Challengerosergia hansjacobi"))

#make shape pallette
shapes.sp <- c("Deosergestes corniculum" = 21,
              "Deosergestes henseni" = 22, 
              "Allosergestes pectinatus" = 23, 
              "Allosergestes sargassi" = 24,
              "Sergestes atlanticus" = 25, 
              "Parasergestes vigilax" = 21, 
              "Parasergestes armatus" = 22,
              "Eusergestes arcticus" = 23, 
              "Gardinerosergia splendens" = 24, 
              "Robustosergia regalis" = 25, 
              "Robustosergia robusta" = 21, 
              "Phorcosergia grandis" = 22,
              "Sergia tenuiremis" = 23,
              "Challengerosergia talismani" = 24,
              "Challengerosergia hansjacobi" = 25)

#rainbow palette
cols.sp <- c(#Sergestes group
              "Deosergestes corniculum" = "#16ABA8",
              "Deosergestes henseni" = "#1f78b4",
              "Allosergestes pectinatus" = "#3acf75",
              "Allosergestes sargassi" = "#33a02c",
              "Sergestes atlanticus" = "#fb9a99",
              "Parasergestes vigilax" = "#e31a1c",
              "Parasergestes armatus" = "#fdbf6f",
              "Eusergestes arcticus" = "#ff7f00",
              #Sergia group
              "Gardinerosergia splendens" = "#E4E4E4",
              "Robustosergia regalis" = "#E4E4E4",
              "Robustosergia robusta" = "#E4E4E4",
              "Phorcosergia grandis" = "#E4E4E4",
              "Sergia tenuiremis" = "#E4E4E4",
              "Challengerosergia talismani" = "#E4E4E4",
              "Challengerosergia hansjacobi" = "#E4E4E4")


#set s limits by species for plots
limits <- specimens_test %>%
  group_by(genus_species) %>%
  summarise(max = max(Body_Length, na.rm = TRUE),
            min = min(Body_Length, na.rm = TRUE))

#make plot
plot_sergestes <- ggplot(filter(specimens_test, Subgroup=="Sergestes"), aes(x = Body_Length, y = Eye_Diameter, color =  species_text, shape = species_text, fill = species_text)) + 
  #Sergia group regressions (in background)
    ##Gardinerosergia splendens
    geom_segment(aes(x = limits$min[which(limits$genus_species == "Gardinerosergia_splendens")], xend = limits$max[which(limits$genus_species == "Gardinerosergia_splendens")], y = 10^(log10(limits$min[which(limits$genus_species == "Gardinerosergia_splendens")])*cc_species["Gardinerosergia_splendens", "slope"] + cc_species["Gardinerosergia_splendens", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Gardinerosergia_splendens")])*cc_species["Gardinerosergia_splendens", "slope"] + cc_species["Gardinerosergia_splendens", "elevation"])), color = cols.sp["Gardinerosergia splendens"]) + 
    ##Robustosergia regalis
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Robustosergia_regalis")], xend = limits$max[which(limits$genus_species == "Robustosergia_regalis")], y = 10^(log10(limits$min[which(limits$genus_species == "Robustosergia_regalis")])*cc_species["Robustosergia_regalis", "slope"] + cc_species["Robustosergia_regalis", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Robustosergia_regalis")])*cc_species["Robustosergia_regalis", "slope"] + cc_species["Robustosergia_regalis", "elevation"])), color = cols.sp["Robustosergia regalis"]) +
    ##Robustosergia robusta
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Robustosergia_robusta")], xend = limits$max[which(limits$genus_species == "Robustosergia_robusta")], y = 10^(log10(limits$min[which(limits$genus_species == "Robustosergia_robusta")])*cc_species["Robustosergia_robusta", "slope"] + cc_species["Robustosergia_robusta", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Robustosergia_robusta")])*cc_species["Robustosergia_robusta", "slope"] + cc_species["Robustosergia_robusta", "elevation"])), color = cols.sp["Robustosergia robusta"]) + 
    ##Phorcosergia grandis
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Phorcosergia_grandis")], xend = limits$max[which(limits$genus_species == "Phorcosergia_grandis")], y = 10^(log10(limits$min[which(limits$genus_species == "Phorcosergia_grandis")])*cc_species["Phorcosergia_grandis", "slope"] + cc_species["Phorcosergia_grandis", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Phorcosergia_grandis")])*cc_species["Phorcosergia_grandis", "slope"] + cc_species["Phorcosergia_grandis", "elevation"])), color = cols.sp["Phorcosergia grandis"]) + 
    ##Sergia tenuiremis
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Sergia_tenuiremis")], xend = limits$max[which(limits$genus_species == "Sergia_tenuiremis")], y = 10^(log10(limits$min[which(limits$genus_species == "Sergia_tenuiremis")])*cc_species["Sergia_tenuiremis", "slope"] + cc_species["Sergia_tenuiremis", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Sergia_tenuiremis")])*cc_species["Sergia_tenuiremis", "slope"] + cc_species["Sergia_tenuiremis", "elevation"])), color = cols.sp["Sergia tenuiremis"])+
    ##Challengosergia talismani
    geom_segment(aes(x = limits$min[which(limits$genus_species == "Challengerosergia_hansjacobi")], xend = limits$max[which(limits$genus_species == "Challengerosergia_hansjacobi")], y = 10^(log10(limits$min[which(limits$genus_species == "Challengerosergia_hansjacobi")])*cc_species["Challengerosergia_hansjacobi", "slope"] + cc_species["Challengerosergia_hansjacobi", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Challengerosergia_hansjacobi")])*cc_species["Challengerosergia_hansjacobi", "slope"] + cc_species["Challengerosergia_hansjacobi", "elevation"])), color = cols.sp["Challengerosergia hansjacobi"]) + 
    ##Challengosergia hansjacobi
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Challengerosergia_talismani")], xend = limits$max[which(limits$genus_species == "Challengerosergia_talismani")], y = 10^(log10(limits$min[which(limits$genus_species == "Challengerosergia_talismani")])*cc_species["Challengerosergia_talismani", "slope"] + cc_species["Challengerosergia_talismani", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Challengerosergia_talismani")])*cc_species["Challengerosergia_talismani", "slope"] + cc_species["Challengerosergia_talismani", "elevation"])), color = cols.sp["Challengerosergia talismani"]) + 
  #Sergestes data points
  geom_point(size = 2, alpha = 0.4) + 
  scale_shape_manual(values = shapes.sp, name = "Species") + 
  scale_color_manual(values = cols.sp, name = "Species") +
  scale_fill_manual(values = cols.sp, name = "Species") +
  scale_x_log10(name = "Body length (mm)", breaks = c(15,20,30,50,100)) + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)", breaks = c(0.2,0.3,0.5,1, 2.6)) + #makes y axis log scale and named
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), legend.text = element_text(face = "italic")) +
  #Sergestes regressions (in forefront)
      ##Allosergestes pectinatus
    geom_segment(aes(x = limits$min[which(limits$genus_species == "Allosergestes_pectinatus")], xend = limits$max[which(limits$genus_species == "Allosergestes_pectinatus")], y = 10^(log10(limits$min[which(limits$genus_species == "Allosergestes_pectinatus")])*cc_species["Allosergestes_pectinatus", "slope"] + cc_species["Allosergestes_pectinatus", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Allosergestes_pectinatus")])*cc_species["Allosergestes_pectinatus", "slope"] + cc_species["Allosergestes_pectinatus", "elevation"])), color = cols.sp["Allosergestes pectinatus"]) +
    ##Allosergestes sargassi
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Allosergestes_sargassi")], xend = limits$max[which(limits$genus_species == "Allosergestes_sargassi")], y = 10^(log10(limits$min[which(limits$genus_species == "Allosergestes_sargassi")])*cc_species["Allosergestes_sargassi", "slope"] + cc_species["Allosergestes_sargassi", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Allosergestes_sargassi")])*cc_species["Allosergestes_sargassi", "slope"] + cc_species["Allosergestes_sargassi", "elevation"])), color = cols.sp["Allosergestes sargassi"]) +
    ##Deosergestes corniculum
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Deosergestes_corniculum")], xend = limits$max[which(limits$genus_species == "Deosergestes_corniculum")], y = 10^(log10(limits$min[which(limits$genus_species == "Deosergestes_corniculum")])*cc_species["Deosergestes_corniculum", "slope"] + cc_species["Deosergestes_corniculum", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Deosergestes_corniculum")])*cc_species["Deosergestes_corniculum", "slope"] + cc_species["Deosergestes_corniculum", "elevation"])), color = cols.sp["Deosergestes corniculum"]) +
    ##Deosergestes henseni
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Deosergestes_henseni")], xend = limits$max[which(limits$genus_species == "Deosergestes_henseni")], y = 10^(log10(limits$min[which(limits$genus_species == "Deosergestes_henseni")])*cc_species["Deosergestes_henseni", "slope"] + cc_species["Deosergestes_henseni", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Deosergestes_henseni")])*cc_species["Deosergestes_henseni", "slope"] + cc_species["Deosergestes_henseni", "elevation"])), color = cols.sp["Deosergestes henseni"]) + 
    ##Parasergestes armatus
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Parasergestes_armatus")], xend = limits$max[which(limits$genus_species == "Parasergestes_armatus")], y = 10^(log10(limits$min[which(limits$genus_species == "Parasergestes_armatus")])*cc_species["Parasergestes_armatus", "slope"] + cc_species["Parasergestes_armatus", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Parasergestes_armatus")])*cc_species["Parasergestes_armatus", "slope"] + cc_species["Parasergestes_armatus", "elevation"])), color = cols.sp["Parasergestes armatus"]) +
    ##Parasergestes vigilax
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Parasergestes_vigilax")], xend = limits$max[which(limits$genus_species == "Parasergestes_vigilax")], y = 10^(log10(limits$min[which(limits$genus_species == "Parasergestes_vigilax")])*cc_species["Parasergestes_vigilax", "slope"] + cc_species["Parasergestes_vigilax", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Parasergestes_vigilax")])*cc_species["Parasergestes_vigilax", "slope"] + cc_species["Parasergestes_vigilax", "elevation"])), color = cols.sp["Parasergestes vigilax"]) +
  ##Sergestes atlanticus
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Sergestes_atlanticus")], xend = limits$max[which(limits$genus_species == "Sergestes_atlanticus")], y = 10^(log10(limits$min[which(limits$genus_species == "Sergestes_atlanticus")])*cc_species["Sergestes_atlanticus", "slope"] + cc_species["Sergestes_atlanticus", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Sergestes_atlanticus")])*cc_species["Sergestes_atlanticus", "slope"] + cc_species["Sergestes_atlanticus", "elevation"])), color = cols.sp["Sergestes atlanticus"]) +
   ##Eusergestes arcticus
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Eusergestes_arcticus")], xend = limits$max[which(limits$genus_species == "Eusergestes_arcticus")], y = 10^(log10(limits$min[which(limits$genus_species == "Eusergestes_arcticus")])*cc_species["Eusergestes_arcticus", "slope"] + cc_species["Eusergestes_arcticus", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Eusergestes_arcticus")])*cc_species["Eusergestes_arcticus", "slope"] + cc_species["Eusergestes_arcticus", "elevation"])), color = cols.sp["Eusergestes arcticus"])+
  theme(legend.text = element_text(face = "italic"))

#interactive plot
ggplotly(plot_sergestes)
```

Then we make a panel showing the -sergia data, with the -sergestes regressions in a light gray for reference. 

```{r species-fig1b, fig.width=9, fig.height=9}

#rainbow palette
cols.sp <- c(#Sergestes group
              "Deosergestes corniculum" = "#E4E4E4",
              "Deosergestes henseni" = "#E4E4E4",
              "Allosergestes pectinatus" = "#E4E4E4",
              "Allosergestes sargassi" = "#E4E4E4",
              "Sergestes atlanticus" = "#E4E4E4",
              "Parasergestes vigilax" = "#E4E4E4",
              "Parasergestes armatus" = "#E4E4E4",
              "Eusergestes arcticus" = "#E4E4E4",
              #Sergia group
              "Gardinerosergia splendens" = "#cab2d6",
              "Robustosergia regalis" = "#6a3d9a",
              "Robustosergia robusta" = "gray31",
              "Phorcosergia grandis" = "burlywood4",
              "Sergia tenuiremis" = "black",
              "Challengerosergia talismani" = "maroon1",
              "Challengerosergia hansjacobi" = "orchid1")

#set s limits by species for plots
limits <- specimens_test %>%
  group_by(genus_species) %>%
  summarise(max = max(Body_Length, na.rm = TRUE),
            min = min(Body_Length, na.rm = TRUE))

#make plot
plot_sergia <- ggplot(filter(specimens_test, Subgroup=="Sergia"), aes(x = Body_Length, y = Eye_Diameter, color =  species_text, shape = species_text, fill = species_text)) + 
    #Sergestes regressions (in background)
      ##Allosergestes pectinatus
    geom_segment(aes(x = limits$min[which(limits$genus_species == "Allosergestes_pectinatus")], xend = limits$max[which(limits$genus_species == "Allosergestes_pectinatus")], y = 10^(log10(limits$min[which(limits$genus_species == "Allosergestes_pectinatus")])*cc_species["Allosergestes_pectinatus", "slope"] + cc_species["Allosergestes_pectinatus", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Allosergestes_pectinatus")])*cc_species["Allosergestes_pectinatus", "slope"] + cc_species["Allosergestes_pectinatus", "elevation"])), color = cols.sp["Allosergestes pectinatus"]) +
    ##Allosergestes sargassi
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Allosergestes_sargassi")], xend = limits$max[which(limits$genus_species == "Allosergestes_sargassi")], y = 10^(log10(limits$min[which(limits$genus_species == "Allosergestes_sargassi")])*cc_species["Allosergestes_sargassi", "slope"] + cc_species["Allosergestes_sargassi", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Allosergestes_sargassi")])*cc_species["Allosergestes_sargassi", "slope"] + cc_species["Allosergestes_sargassi", "elevation"])), color = cols.sp["Allosergestes sargassi"]) +
    ##Deosergestes corniculum
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Deosergestes_corniculum")], xend = limits$max[which(limits$genus_species == "Deosergestes_corniculum")], y = 10^(log10(limits$min[which(limits$genus_species == "Deosergestes_corniculum")])*cc_species["Deosergestes_corniculum", "slope"] + cc_species["Deosergestes_corniculum", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Deosergestes_corniculum")])*cc_species["Deosergestes_corniculum", "slope"] + cc_species["Deosergestes_corniculum", "elevation"])), color = cols.sp["Deosergestes corniculum"]) +
    ##Deosergestes henseni
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Deosergestes_henseni")], xend = limits$max[which(limits$genus_species == "Deosergestes_henseni")], y = 10^(log10(limits$min[which(limits$genus_species == "Deosergestes_henseni")])*cc_species["Deosergestes_henseni", "slope"] + cc_species["Deosergestes_henseni", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Deosergestes_henseni")])*cc_species["Deosergestes_henseni", "slope"] + cc_species["Deosergestes_henseni", "elevation"])), color = cols.sp["Deosergestes henseni"]) + 
    ##Parasergestes armatus
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Parasergestes_armatus")], xend = limits$max[which(limits$genus_species == "Parasergestes_armatus")], y = 10^(log10(limits$min[which(limits$genus_species == "Parasergestes_armatus")])*cc_species["Parasergestes_armatus", "slope"] + cc_species["Parasergestes_armatus", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Parasergestes_armatus")])*cc_species["Parasergestes_armatus", "slope"] + cc_species["Parasergestes_armatus", "elevation"])), color = cols.sp["Parasergestes armatus"]) +
    ##Parasergestes vigilax
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Parasergestes_vigilax")], xend = limits$max[which(limits$genus_species == "Parasergestes_vigilax")], y = 10^(log10(limits$min[which(limits$genus_species == "Parasergestes_vigilax")])*cc_species["Parasergestes_vigilax", "slope"] + cc_species["Parasergestes_vigilax", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Parasergestes_vigilax")])*cc_species["Parasergestes_vigilax", "slope"] + cc_species["Parasergestes_vigilax", "elevation"])), color = cols.sp["Parasergestes vigilax"]) +
  ##Sergestes atlanticus
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Sergestes_atlanticus")], xend = limits$max[which(limits$genus_species == "Sergestes_atlanticus")], y = 10^(log10(limits$min[which(limits$genus_species == "Sergestes_atlanticus")])*cc_species["Sergestes_atlanticus", "slope"] + cc_species["Sergestes_atlanticus", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Sergestes_atlanticus")])*cc_species["Sergestes_atlanticus", "slope"] + cc_species["Sergestes_atlanticus", "elevation"])), color = cols.sp["Sergestes atlanticus"]) +
   ##Eusergestes arcticus
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Eusergestes_arcticus")], xend = limits$max[which(limits$genus_species == "Eusergestes_arcticus")], y = 10^(log10(limits$min[which(limits$genus_species == "Eusergestes_arcticus")])*cc_species["Eusergestes_arcticus", "slope"] + cc_species["Eusergestes_arcticus", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Eusergestes_arcticus")])*cc_species["Eusergestes_arcticus", "slope"] + cc_species["Eusergestes_arcticus", "elevation"])), color = cols.sp["Eusergestes arcticus"])+
  theme(legend.text = element_text(face = "italic"))+
  #Sergia data points
  geom_point(size = 2, alpha = 0.4) + 
  scale_shape_manual(values = shapes.sp, name = "Species") + 
  scale_color_manual(values = cols.sp, name = "Species") +
  scale_fill_manual(values = cols.sp, name = "Species") +
  scale_x_log10(name = "Body length (mm)", breaks = c(15,20,30,50,100)) + #makes x axis log scale and named
  scale_y_log10(name = "Eye diameter (mm)", breaks = c(0.2,0.3,0.5,1, 2.6)) + #makes y axis log scale and named
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), legend.text = element_text(face = "italic")) +
  #Sergia group regressions (in forefront)
    ##Gardinerosergia splendens
    geom_segment(aes(x = limits$min[which(limits$genus_species == "Gardinerosergia_splendens")], xend = limits$max[which(limits$genus_species == "Gardinerosergia_splendens")], y = 10^(log10(limits$min[which(limits$genus_species == "Gardinerosergia_splendens")])*cc_species["Gardinerosergia_splendens", "slope"] + cc_species["Gardinerosergia_splendens", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Gardinerosergia_splendens")])*cc_species["Gardinerosergia_splendens", "slope"] + cc_species["Gardinerosergia_splendens", "elevation"])), color = cols.sp["Gardinerosergia splendens"]) + 
    ##Robustosergia regalis
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Robustosergia_regalis")], xend = limits$max[which(limits$genus_species == "Robustosergia_regalis")], y = 10^(log10(limits$min[which(limits$genus_species == "Robustosergia_regalis")])*cc_species["Robustosergia_regalis", "slope"] + cc_species["Robustosergia_regalis", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Robustosergia_regalis")])*cc_species["Robustosergia_regalis", "slope"] + cc_species["Robustosergia_regalis", "elevation"])), color = cols.sp["Robustosergia regalis"]) +
    ##Robustosergia robusta
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Robustosergia_robusta")], xend = limits$max[which(limits$genus_species == "Robustosergia_robusta")], y = 10^(log10(limits$min[which(limits$genus_species == "Robustosergia_robusta")])*cc_species["Robustosergia_robusta", "slope"] + cc_species["Robustosergia_robusta", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Robustosergia_robusta")])*cc_species["Robustosergia_robusta", "slope"] + cc_species["Robustosergia_robusta", "elevation"])), color = cols.sp["Robustosergia robusta"]) + 
    ##Phorcosergia grandis
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Phorcosergia_grandis")], xend = limits$max[which(limits$genus_species == "Phorcosergia_grandis")], y = 10^(log10(limits$min[which(limits$genus_species == "Phorcosergia_grandis")])*cc_species["Phorcosergia_grandis", "slope"] + cc_species["Phorcosergia_grandis", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Phorcosergia_grandis")])*cc_species["Phorcosergia_grandis", "slope"] + cc_species["Phorcosergia_grandis", "elevation"])), color = cols.sp["Phorcosergia grandis"]) + 
    ##Sergia tenuiremis
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Sergia_tenuiremis")], xend = limits$max[which(limits$genus_species == "Sergia_tenuiremis")], y = 10^(log10(limits$min[which(limits$genus_species == "Sergia_tenuiremis")])*cc_species["Sergia_tenuiremis", "slope"] + cc_species["Sergia_tenuiremis", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Sergia_tenuiremis")])*cc_species["Sergia_tenuiremis", "slope"] + cc_species["Sergia_tenuiremis", "elevation"])), color = cols.sp["Sergia tenuiremis"])+
    ##Challengosergia talismani
    geom_segment(aes(x = limits$min[which(limits$genus_species == "Challengerosergia_hansjacobi")], xend = limits$max[which(limits$genus_species == "Challengerosergia_hansjacobi")], y = 10^(log10(limits$min[which(limits$genus_species == "Challengerosergia_hansjacobi")])*cc_species["Challengerosergia_hansjacobi", "slope"] + cc_species["Challengerosergia_hansjacobi", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Challengerosergia_hansjacobi")])*cc_species["Challengerosergia_hansjacobi", "slope"] + cc_species["Challengerosergia_hansjacobi", "elevation"])), color = cols.sp["Challengerosergia hansjacobi"]) + 
    ##Challengosergia hansjacobi
  geom_segment(aes(x = limits$min[which(limits$genus_species == "Challengerosergia_talismani")], xend = limits$max[which(limits$genus_species == "Challengerosergia_talismani")], y = 10^(log10(limits$min[which(limits$genus_species == "Challengerosergia_talismani")])*cc_species["Challengerosergia_talismani", "slope"] + cc_species["Challengerosergia_talismani", "elevation"]), yend = 10^(log10(limits$max[which(limits$genus_species == "Challengerosergia_talismani")])*cc_species["Challengerosergia_talismani", "slope"] + cc_species["Challengerosergia_talismani", "elevation"])), color = cols.sp["Challengerosergia talismani"])

#interactive plot
ggplotly(plot_sergia)
```


```{r, results = "hide"}

# Build legend for sergia----

#cols and shapes
cols.sergia <- c("Gardinerosergia splendens" = "#cab2d6",
              "Robustosergia regalis" = "#6a3d9a",
              "Robustosergia robusta" = "gray31",
              "Phorcosergia grandis" = "burlywood4",
              "Sergia tenuiremis" = "black",
              "Challengerosergia talismani" = "maroon1",
              "Challengerosergia hansjacobi" = "orchid1")

shapes.sergia <- c("Gardinerosergia splendens" = 24, 
              "Robustosergia regalis" = 25, 
              "Robustosergia robusta" = 21, 
              "Phorcosergia grandis" = 22,
              "Sergia tenuiremis" = 23,
              "Challengerosergia talismani" = 24,
              "Challengerosergia hansjacobi" = 25)

#make -sergia legend with alpha = 1
plot_sergia_leg <- ggplot(filter(specimens_test, Subgroup=="Sergia"), aes(x = Body_Length, y = Eye_Diameter, color =  species_text, shape = species_text, fill = species_text)) +
  geom_point(size = 2, alpha = 1) + 
  scale_shape_manual(values = shapes.sergia, name = "-sergia") + 
  scale_color_manual(values = cols.sergia, name = "-sergia") +
  scale_fill_manual(values = cols.sergia, name = "-sergia") +
  theme_bw() +
  theme(legend.text = element_text(face = "italic"), legend.title = element_text(face = "italic"))

# extract legend from this plot with alpha = 1
leg.sergia <- get_legend(plot_sergia_leg)

# Build legend for sergestes-----

#cols and shapes
cols.sergestes <- c("Deosergestes corniculum" = "#16ABA8",
              "Deosergestes henseni" = "#1f78b4",
              "Allosergestes pectinatus" = "#3acf75",
              "Allosergestes sargassi" = "#33a02c",
              "Sergestes atlanticus" = "#fb9a99",
              "Parasergestes vigilax" = "#e31a1c",
              "Parasergestes armatus" = "#fdbf6f",
              "Eusergestes arcticus" = "#ff7f00")

shapes.sergestes <- c("Deosergestes corniculum" = 21,
              "Deosergestes henseni" = 22, 
              "Allosergestes pectinatus" = 23, 
              "Allosergestes sargassi" = 24,
              "Sergestes atlanticus" = 25, 
              "Parasergestes vigilax" = 21, 
              "Parasergestes armatus" = 22,
              "Eusergestes arcticus" = 23)

labs.sergestes <- c("Deosergestes corniculum",
              "Deosergestes henseni", 
              "Allosergestes pectinatus *", 
              "Allosergestes sargassi *",
              "Sergestes atlanticus *", 
              "Parasergestes vigilax *", 
              "Parasergestes armatus",
              "Eusergestes arcticus")

#make -sergia legend with alpha = 1
plot_sergestes_leg <- ggplot(filter(specimens_test, Subgroup=="Sergestes"), aes(x = Body_Length, y = Eye_Diameter, color =  species_text, shape = species_text, fill = species_text)) +
  geom_point(size = 2, alpha = 1) + 
  scale_shape_manual(values = shapes.sergestes, name = "-sergestes", 
                     labels = labs.sergestes) + 
  scale_color_manual(values = cols.sergestes, name = "-sergestes", 
                     labels = labs.sergestes) + 
  scale_fill_manual(values = cols.sergestes, name = "-sergestes", 
                     labels = labs.sergestes) + 
  theme_bw() +
  theme(legend.text = element_text(face = "italic"), legend.title = element_text(face = "italic"))

# extract legend from this plot with alpha = 1
leg.sergestes <- get_legend(plot_sergestes_leg)

```

```{r}

fig.a <- plot_sergestes + 
  theme_classic() +
  theme(legend.position = "none") 


fig.b <- plot_sergia +
  theme_classic() +
  theme(legend.position = "none", axis.title.y = element_blank(), axis.text.y = element_blank())
  

# arrange panels in figure
plots <- plot_grid(fig.a, fig.b,
           align = 'vh', #align horizontally and vertically
           axis = 'lb',
           labels = c("A", "B"),
           hjust = -0.5, #adjustment for panel labels
           nrow = 1) #number of rows in grids

#arrange legends vertically
legends <- plot_grid(leg.sergestes, leg.sergia,
                     align = 'vh', 
                     axis = 't',
                     nrow = 2) #number of rows in grids

#put together figure
plot_grid(plots, legends, ncol = 2, rel_widths = c(1, 0.5), align = 'vh')

#export figure
pdf("../Figures/Fig-1-revised.pdf", width = 12, height = 4.5)
plot_grid(plots, legends, ncol = 2, rel_widths = c(1, 0.5), align = 'vh')
dev.off()

#eps format
setEPS()
postscript("../Figures/Fig-1-revised.eps", width = 12, height = 4.5)
plot_grid(plots, legends, ncol = 2, rel_widths = c(1, 0.5), align = 'vh')
dev.off()
```
